requirements/common.txt
Django==4.1.*
Pillow==9.4.*
python-dotenv==1.0.0
